Refactor: change project structure (#138)
* Refactor: change tests package name to run

* Refactor: delete suite folder
